src/aoai-api-simulator/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0.post1
gunicorn==21.2.0
requests==2.31.0
PyYAML==6.0.1
tiktoken==0.6.0
nanoid==2.0.0
limits==3.8.0
azure-monitor-opentelemetry==1.3.0
pydantic-settings==2.2.1
